+++ core-data/src/commonMain/kotlin/ai/saniou/thread/data/source/tieba/model/LikeForumResultBean.kt
@@ -8,37 +8,37 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class LikeForumResultBean(
     @SerialName("error_code")
-    val errorCode: String,
-    val error: ErrorInfo,
-    val info: Info,
-    val userPerm: UserPermInfo
+    val errorCode: String = "0",
+    val error: ErrorInfo? = null,
+    val info: Info? = null,
+    val userPerm: UserPermInfo? = null,
 ) {
     @Serializable
     data class UserPermInfo(
         @SerialName("level_id")
-        val levelId: String,
+        val levelId: String = "",
         @SerialName("level_name")
-        val levelName: String
+        val levelName: String = "",
     )
 
     @Serializable
     data class Info(
         @SerialName("cur_score")
-        val curScore: String,
+        val curScore: String = "",
         @SerialName("levelup_score")
-        val levelUpScore: String,
+        val levelUpScore: String = "",
         @SerialName("level_id")
-        val levelId: String,
+        val levelId: String = "",
         @SerialName("level_name")
-        val levelName: String,
+        val levelName: String = "",
         @SerialName("member_sum")
-        val memberSum: String
+        val memberSum: String = "",
     )
 
     @Serializable
     data class ErrorInfo(
-        val errno: String,
-        val errmsg: String,
-        val usermsg: String
+        val errno: String = "",
+        val errmsg: String = "",
+        val usermsg: String = "",
     )
 }